backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas
numpy
pyarrow>=10.0.0
duckdb>=0.9.0
python-dotenv==1.0.0
langchain==0.3.27
langchain-aws>=0.1.0
langchain-core==0.3.76
langchain-community==0.3.29
langchain-text-splitters==0.3.11
boto3>=1.34.0
networkx